services/kb-service/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Embeddings + Vector DB
sentence-transformers==3.0.1
faiss-gpu==1.7.2

# Utils
numpy==1.26.4
pydantic==2.9.2
